debut de la refactorisation de relationship
