+++ elrahapi/relationship/base_relationship.py
@@ -1,44 +1,23 @@
 from copy import deepcopy
 from typing import Any, Type
 
+from pydantic import BaseModel
 from elrahapi.authentication.authentication_manager import AuthenticationManager
 from elrahapi.crud.crud_forgery import CrudForgery
-from elrahapi.crud.crud_models import CrudModels
-from elrahapi.exception.exceptions_utils import raise_custom_http_exception
-from elrahapi.router.route_config import (
-    AuthorizationConfig,
-    ResponseModelConfig,
-    RouteConfig,
-)
-from elrahapi.router.router_crud import (
-    add_authorizations,
-    initialize_dependecies,
-    is_verified_relation_rule,
-    set_response_models,
-)
-from elrahapi.router.router_namespace import TypeRelation
+from elrahapi.router.route_additional_config import AuthorizationConfig, ResponseModelConfig
+from elrahapi.router.route_config import RouteConfig
+from elrahapi.router.router_crud import add_authorizations, initialize_dependecies, is_verified_relation_rule, set_response_models
 from elrahapi.router.router_routes_name import RelationRoutesName
-from elrahapi.utility.types import ElrahSession
-from elrahapi.utility.utils import apply_filters, exec_stmt, validate_value
-from pydantic import BaseModel
-from sqlalchemy import and_, select
-from sqlalchemy.sql.schema import Table
-
-from fastapi import status
+from elrahapi.utility.utils import validate_value
 
 
-class Relationship:
+class BaseRelationship:
 
     def __init__(
         self,
         relationship_name: str,
-        type_relation: TypeRelation,
         second_entity_crud: CrudForgery,
-        relationship_crud: CrudForgery | None = None,
-        relationship_key1_name: str | None = None,
-        relationship_key2_name: str | None = None,
         relations_routes_configs: list[RouteConfig] | None = None,
-        relation_table: Table | None = None,
         second_entity_fk_name: str | None = None,
         relations_authorizations_configs: AuthorizationConfig | None = None,
         relations_responses_model_configs: ResponseModelConfig | None = None,
@@ -48,12 +27,7 @@ class Relationship:
         self.relationship_name = relationship_name
         self.second_entity_fk_name = second_entity_fk_name
         self.relationship_name = relationship_name
-        self.relationship_crud = relationship_crud
         self.second_entity_crud = second_entity_crud
-        self.relationship_key1_name = relationship_key1_name
-        self.relationship_key2_name = relationship_key2_name
-        self.type_relation = type_relation
-        self.relation_table = relation_table
         self.relations_routes_configs = relations_routes_configs or []
         self.default_public_relation_routes_name = (
             default_public_relation_routes_name or []
@@ -63,30 +37,16 @@ class Relationship:
         )
         self.relations_authorizations_configs = relations_authorizations_configs or []
         self.relations_responses_model_configs = relations_responses_model_configs or []
-        self.check_relation_rules()
 
     def get_second_model_key(self):
         return self.second_entity_crud.crud_models.get_pk()
 
-    def get_relationship_keys(self):
-        if self.type_relation == TypeRelation.MANY_TO_MANY_CLASS:
-            rel_key1 = self.relationship_crud.crud_models.get_attr(
-                self.relationship_key1_name
-            )
-            rel_key2 = self.relationship_crud.crud_models.get_attr(
-                self.relationship_key2_name
-            )
-            return rel_key1, rel_key2
-        elif self.type_relation == TypeRelation.MANY_TO_MANY_TABLE:
-            if self.relation_table is not None:
-                columns = self.relation_table.c
-                rel_key1 = getattr(columns, self.relationship_key1_name)
-                rel_key2 = getattr(columns, self.relationship_key2_name)
-                return rel_key1, rel_key2
-        else:
-            raise ValueError(
-                f"relationship_keys not available for relation type {self.type_relation}"
-            )
+    def add_fk(self, obj: Type[BaseModel], fk: Any):
+        if self.second_entity_fk_name is not None:
+            validated_fk = validate_value(value=fk)
+            new_obj = obj.model_copy(update={self.second_entity_fk_name: validated_fk})
+            return new_obj
+        return obj
 
     def check_relation_rules(self):
         for route_config in self.relations_routes_configs:
@@ -103,6 +63,7 @@ class Relationship:
         default_public_relation_routes_name: list[RelationRoutesName],
         default_protected_relation_routes_name: list[RelationRoutesName],
     ):
+        self.check_relation_rules()
         full_routes_configs = (
             default_public_relation_routes_name + default_protected_relation_routes_name
         )
@@ -324,233 +285,3 @@ class Relationship:
                     privileges=privileges,
                 )
         return routes_configs
-
-    async def create_relation(
-        self, session: ElrahSession, entity_crud: CrudForgery, pk1: Any, pk2: Any
-    ):
-        entity_1 = await entity_crud.read_one(session=session, pk=pk1)
-        entity_2 = await self.second_entity_crud.read_one(session=session, pk=pk2)
-        if self.type_relation == TypeRelation.ONE_TO_ONE:
-            setattr(entity_1, self.relationship_name, entity_2)
-        elif self.type_relation in [
-            TypeRelation.MANY_TO_MANY_TABLE,
-            TypeRelation.ONE_TO_MANY,
-        ]:
-            entity_1_attr = getattr(entity_1, self.relationship_name)
-            entity_1_attr.append(entity_2)
-        await entity_crud.session_manager.commit_and_refresh(
-            session=session, object=entity_1
-        )
-
-    async def delete_relation(
-        self, session: ElrahSession, entity_crud: CrudForgery, pk1: Any, pk2: Any
-    ):
-        entity_1 = await entity_crud.read_one(session=session, pk=pk1)
-        entity_2 = await self.second_entity_crud.read_one(session=session, pk=pk2)
-
-        if self.type_relation == TypeRelation.ONE_TO_ONE:
-            setattr(entity_1, self.relationship_name, None)
-            await entity_crud.session_manager.commit_and_refresh(
-                session=session, object=entity_1
-            )
-        elif self.type_relation in [
-            TypeRelation.MANY_TO_MANY_TABLE,
-            TypeRelation.ONE_TO_MANY,
-        ]:
-            entity_1_attr = getattr(entity_1, self.relationship_name)
-            if entity_2 in entity_1_attr:
-                entity_1_attr.remove(entity_2)
-                await entity_crud.session_manager.commit_and_refresh(
-                    session=session, object=entity_1
-                )
-            else:
-                detail = f"Relation between {entity_crud.entity_name} with pk {pk1} and {self.second_entity_crud.entity_name} with pk {pk2} not found"
-                raise_custom_http_exception(
-                    status_code=status.HTTP_404_NOT_FOUND, detail=detail
-                )
-        elif self.type_relation == TypeRelation.MANY_TO_MANY_CLASS:
-            rel = await self.read_one_relation(session=session, pk1=pk1, pk2=pk2)
-            rel_pk = getattr(rel, self.relationship_crud.primary_key_name)
-            return await self.relationship_crud.delete(session=session, pk=rel_pk)
-
-    async def read_one_relation(self, session: ElrahSession, pk1: Any, pk2: Any):
-        rel_key1, rel_key2 = self.get_relationship_keys()
-        if self.type_relation == TypeRelation.MANY_TO_MANY_CLASS:
-            stmt = select(self.relationship_crud.crud_models.sqlalchemy_model).where(
-                and_(rel_key1 == pk1, rel_key2 == pk2)
-            )
-            result = await exec_stmt(
-                stmt=stmt,
-                session=session,
-            )
-            rel = result.scalar_one_or_none()
-            if rel is None:
-                detail = f"Relation of {self.relationship_name} with IDs ({pk1},{pk2}) is not found "
-                raise_custom_http_exception(
-                    status_code=status.HTTP_404_NOT_FOUND, detail=detail
-                )
-            return rel
-        else:
-            detail = f"Bad Operation for relation {self.type_relation} of relationship {self.relationship_name}"
-            raise_custom_http_exception(
-                status_code=status.HTTP_400_BAD_REQUEST, detail=detail
-            )
-
-    def add_fk(self, obj: Type[BaseModel], fk: Any):
-        if self.second_entity_fk_name is not None:
-            validated_fk = validate_value(value=fk)
-            new_obj = obj.model_copy(update={self.second_entity_fk_name: validated_fk})
-            return new_obj
-        return obj
-
-    async def create_by_relation(
-        self,
-        session: ElrahSession,
-        pk1: Any,
-        create_obj: Type[BaseModel],
-        entity_crud: CrudForgery,
-    ):
-        e1 = await entity_crud.read_one(session=session, pk=pk1)
-        e2 = getattr(e1, self.relationship_name)
-        if self.type_relation == TypeRelation.ONE_TO_ONE and e2 is not None:
-            detail = f"{self.second_entity_crud.entity_name} already exists for {entity_crud.entity_name} with pk {pk1}"
-            raise_custom_http_exception(
-                status_code=status.HTTP_400_BAD_REQUEST, detail=detail
-            )
-        create_obj = self.add_fk(obj=create_obj, fk=pk1)
-        new_obj = await self.second_entity_crud.create(
-            session=session, create_obj=create_obj
-        )
-        pk2 = getattr(new_obj, self.second_entity_crud.primary_key_name)
-        await self.create_relation(
-            session=session, entity_crud=entity_crud, pk1=pk1, pk2=pk2
-        )
-        return new_obj
-
-    async def delete_by_relation(
-        self, session: ElrahSession, pk1: Any, entity_crud: CrudForgery
-    ):
-        entity_1 = await entity_crud.read_one(session=session, pk=pk1)
-        entity_2 = getattr(entity_1, self.relationship_name)
-        e2_pk = getattr(entity_2, self.second_entity_crud.primary_key_name)
-        entity_2 = None
-        return await self.second_entity_crud.delete(session=session, pk=e2_pk)
-
-    async def soft_delete_by_relation(
-        self, session: ElrahSession, pk1: Any, entity_crud: CrudForgery
-    ):
-        entity_1 = await entity_crud.read_one(session=session, pk=pk1)
-        entity_2 = getattr(entity_1, self.relationship_name)
-        e2_pk = getattr(entity_2, self.second_entity_crud.primary_key_name)
-        entity_2 = None
-        return await self.second_entity_crud.soft_delete(session=session, pk=e2_pk)
-
-    async def read_one_by_relation(
-        self, session: ElrahSession, pk1: Any, entity_crud: CrudForgery
-    ):
-        e1 = await entity_crud.read_one(session=session, pk=pk1)
-        e2 = getattr(e1, self.relationship_name)
-        if e2 is None:
-            detail = f"{self.relationship_name} not found for {entity_crud.entity_name} with pk {pk1}"
-            raise_custom_http_exception(
-                status_code=status.HTTP_404_NOT_FOUND, detail=detail
-            )
-        return e2
-
-    async def update_by_relation(
-        self,
-        pk1: Any,
-        session: ElrahSession,
-        update_obj: Type[BaseModel],
-        entity_crud: CrudForgery,
-    ):
-        update_obj = self.add_fk(obj=update_obj, fk=pk1)
-        entity = await entity_crud.read_one(session=session, pk=pk1)
-        entity_2 = getattr(entity, self.relationship_name)
-        pk2 = getattr(entity_2, self.second_entity_crud.primary_key_name)
-        return await self.second_entity_crud.update(
-            session=session, pk=pk2, update_obj=update_obj, is_full_update=True
-        )
-
-    async def patch_by_relation(
-        self,
-        session: ElrahSession,
-        pk1: Any,
-        patch_obj: Type[BaseModel],
-        entity_crud: CrudForgery,
-    ):
-        update_obj = self.add_fk(obj=update_obj, fk=pk1)
-        entity = await entity_crud.read_one(session=session, pk=pk1)
-        entity_2 = getattr(entity, self.relationship_name)
-        pk2 = getattr(entity_2, self.second_entity_crud.primary_key_name)
-        return await self.second_entity_crud.update(
-            session=session, pk=pk2, update_obj=patch_obj, is_full_update=False
-        )
-
-    async def read_all_by_relation(
-        self,
-        session: ElrahSession,
-        entity_crud: CrudForgery,
-        pk1: Any,
-        skip: int = 0,
-        limit: int = None,
-        filters=dict[str, Any],
-    ):
-        e2_cm: CrudModels = self.second_entity_crud.crud_models
-        e1_cm = entity_crud.crud_models
-        e2_pk = e2_cm.get_pk()
-        e1_pk = e1_cm.get_pk()
-        if self.type_relation == TypeRelation.ONE_TO_MANY:
-            stmt = (
-                select(e2_cm.sqlalchemy_model)
-                .join(e1_cm.sqlalchemy_model)
-                .where(e1_pk == pk1)
-            )
-            stmt = apply_filters(crud_models=e2_cm, stmt=stmt, filters=filters)
-            stmt = stmt.offset(skip).limit(limit)
-            results = await exec_stmt(
-                session=session,
-                with_scalars=True,
-                stmt=stmt,
-            )
-            return results.all()
-        elif self.type_relation == TypeRelation.MANY_TO_MANY_CLASS:
-            rel_model = self.relationship_crud.crud_models.sqlalchemy_model
-            rel_key1, rel_key2 = self.get_relationship_keys()
-            stmt = select(e2_cm.sqlalchemy_model).join(rel_model).where(rel_key1 == pk1)
-            stmt = apply_filters(
-                crud_models=e2_cm,
-                stmt=stmt,
-                filters=filters,
-            )
-            stmt = stmt.offset(skip).limit(limit)
-            results = await exec_stmt(
-                session=session,
-                with_scalars=True,
-                stmt=stmt,
-            )
-            return results.all()
-        elif self.type_relation == TypeRelation.MANY_TO_MANY_TABLE:
-            rel_key1, rel_key2 = self.get_relationship_keys()
-            stmt = (
-                select(e2_cm.sqlalchemy_model)
-                .join(self.relation_table, e2_pk == rel_key2)
-                .join(e1_cm.sqlalchemy_model, rel_key1 == e1_pk)
-            )
-            stmt = apply_filters(
-                crud_models=e2_cm,
-                stmt=stmt,
-                filters=filters,
-            )
-            stmt = stmt.offset(skip).limit(limit)
-            results = await exec_stmt(
-                session=session,
-                with_scalars=True,
-                stmt=stmt,
-            )
-            return results.all()
-        else:
-            detail = f"Operation Invalid for relation {self.type_relation}"
-            raise_custom_http_exception(
-                status_code=status.HTTP_400_BAD_REQUEST, detail=detail
-            )
